day3_agent/requirements.txt
google-generativeai
pydantic
python-dotenv
groq
